change transaction

+++ database/migrations/2023_07_25_115046_add_balance_to_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->float('balance')->default(0)->after('role');
+            $table->double('balance', 15, 5)->default(0)->after('role');
         });
     }
